+++ NEWS.md
@@ -3,6 +3,7 @@
 * The `test` argument of `compare_sequences()` is now `TRUE` by default for backward compatibility.
 * Fixed an issue with `bootstrap()` when the model contained weights that were zero.
 * Fixed issues with several distance measures in `cluster_sequences()`.
+* Fixed edge ordering in `betweenness_network()`.
 
 # tna 1.2.0