+++ NEWS.md
@@ -1,4 +1,7 @@
 # tna 1.2.1
+* Fix bug in `plot.tna_sequence_comparison`
+
+# tna 1.2.1
 
 * The `test` argument of `compare_sequences()` is now `TRUE` by default for backward compatibility.
 * Fixed an issue with `bootstrap()` when the model contained weights that were zero.